+++ srcjs/amChart4.jsx
@@ -6553,9 +6553,6 @@ class AmPercentageBarChart extends React.PureComponent {
       if (shinyId === undefined) {
         shinyId = $(document.getElementById(chartId)).parent().attr("id");
       }
-      Shiny.setInputValue(
-        shinyId + ":rAmCharts4.dataframe", dataCopy
-      );
     }
 
     switch (theme) {
@@ -6679,6 +6676,60 @@ class AmPercentageBarChart extends React.PureComponent {
           x.column.isHover = false;
         })
       });
+    }
+
+
+    /* ~~~~\  Shiny message handler for percentage bar chart  /~~~~ */
+    if (window.Shiny) {
+      Shiny.addCustomMessageHandler(
+        shinyId + "percentage",
+        function (newdata) {
+          let tail = " is missing in the data you supplied!";
+          // check that the received data has the 'category' column
+          if (!newdata.hasOwnProperty(category)) {
+            console.warn(
+              `updateAmPercentageBarChart: column "${category}"` + tail
+            );
+            return null;
+          }
+          // check that the received data has the necessary categories
+          let ok = true, i = 0;
+          while (ok && i < categories.length) {
+            ok = newdata[category].indexOf(categories[i]) > -1;
+            if (!ok) {
+              console.warn(
+                `updateAmPercentageBarChart: category "${categories[i]}"` + tail
+              );
+            }
+            i++;
+          }
+          if (!ok) {
+            return null;
+          }
+          // check that the received data has the necessary 'values' columns
+          i = 0;
+          while (ok && i < values.length) {
+            ok = newdata.hasOwnProperty(values[i]);
+            if (!ok) {
+              console.warn(
+                `updateAmPercentageBarChart: column "${values[i]}"` + tail
+              );
+            }
+            i++;
+          }
+          if (!ok) {
+            return null;
+          }
+          // update chart data
+          let tnewdata = HTMLWidgets.dataframeToD3(newdata);
+          for (let r = 0; r < data.length; ++r) {
+            for (let v = 0; v < values.length; ++v) {
+              chart.data[r][values[v]] = tnewdata[r][values[v]];
+            }
+          }
+          chart.invalidateRawData();
+        }
+      );
     }